Move trade verifier and add trait
