+++ crates/shared/src/price_estimation/native.rs
@@ -100,6 +100,7 @@ mod tests {
                     out_amount: 123_456_789_000_000_000u128.into(),
                     gas: 0,
                     solver: H160([1; 20]),
+                    verified: false,
                 })
             }
             .boxed()
