+++ crates/shared/src/price_estimation/zeroex.rs
@@ -1,6 +1,7 @@
 use {
     super::{
-        trade_finder::{TradeEstimator, TradeVerifier},
+        trade_finder::TradeEstimator,
+        trade_verifier::TradeVerifying,
         PriceEstimateResult,
         PriceEstimating,
         Query,
@@ -33,7 +34,7 @@ impl ZeroExPriceEstimator {
         ))
     }
 
-    pub fn verified(&self, verifier: TradeVerifier) -> Self {
+    pub fn verified(&self, verifier: Arc<dyn TradeVerifying>) -> Self {
         Self(self.0.clone().with_verifier(verifier))
     }
 }
